# perf: faster `version.sh`