@domoritz 2020-02-14
Nice! I've been bothered by the runtime of the version script for a while and this makes things so much nicer. Thank you so much @axiomabsolute!