@domoritz 2020-02-14
@axiomabsolute Just checking, did you enable github actions on your fork? I wonder how we can make sure that tests run on forks. Maybe we need https://github.com/vega/vega-lite/pull/5910.