@domoritz 2020-02-14
Looks like we need to also run on pull_request but then we can't push updates. I don't see an easy way to support our workflow right now.